Small modifications made
